+++ src/query/storages/common/cache/src/caches.rs
@@ -17,24 +17,9 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use arrow::array::ArrayRef;
-use bytes::Bytes;
 use databend_common_cache::MemSized;
-use databend_common_catalog::plan::PartStatistics;
-use databend_common_catalog::plan::Partitions;
-use databend_common_exception::ErrorCode;
-use databend_common_catalog::table::Table;
-use databend_storages_common_index::filters::Xor8Filter;
-use databend_storages_common_index::BloomIndexMeta;
-use databend_storages_common_index::InvertedIndexFile;
-use databend_storages_common_index::InvertedIndexMeta;
-use databend_storages_common_table_meta::meta::column_oriented_segment::ColumnOrientedSegment;
-use databend_storages_common_table_meta::meta::BlockMeta;
-use databend_storages_common_table_meta::meta::CompactSegmentInfo;
-use databend_storages_common_table_meta::meta::SegmentInfo;
-use databend_storages_common_table_meta::meta::TableSnapshot;
-use databend_storages_common_table_meta::meta::TableSnapshotStatistics;
-use parquet::file::metadata::ParquetMetaData;
 
+use crate::cache_items::*;
 use crate::manager::CacheManager;
 use crate::providers::HybridCache;
 use crate::CacheAccessor;
@@ -42,41 +27,6 @@ use crate::InMemoryLruCache;
 
 /// In memory object cache of SegmentInfo
 pub type CompactSegmentInfoCache = InMemoryLruCache<CompactSegmentInfo>;
-
-// TODO move these out
-pub struct ColumnData(Bytes);
-
-impl ColumnData {
-    pub fn new(bytes: Bytes) -> Self {
-        // TODO review it, data may refer to a shared buffer, which is large.
-        ColumnData(bytes)
-    }
-
-    pub fn bytes(&self) -> Bytes {
-        self.0.clone()
-    }
-
-    pub fn size(&self) -> usize {
-        self.0.len()
-    }
-}
-
-pub type ColumnDataCache = HybridCache<ColumnData>;
-impl TryFrom<&ColumnData> for Vec<u8> {
-    type Error = ErrorCode;
-    fn try_from(value: &ColumnData) -> Result<Self, Self::Error> {
-        Ok(value.0.to_vec())
-    }
-}
-
-// copying should be avoided
-impl TryFrom<Bytes> for ColumnData {
-    type Error = ErrorCode;
-
-    fn try_from(value: Bytes) -> Result<Self, Self::Error> {
-        Ok(ColumnData(value))
-    }
-}
 
 /// In memory object cache of ColumnOrientedSegmentInfo
 pub type ColumnOrientedSegmentInfoCache = InMemoryLruCache<ColumnOrientedSegment>;
