+++ src/query/storages/common/cache/src/read/readers.rs
@@ -14,10 +14,9 @@
 
 use crate::providers::HybridCache;
 use crate::read::cached_reader::CacheReader;
+use crate::read::cached_ttl_reader::CacheTTLReader;
 use crate::InMemoryLruCache;
-use crate::read::cached_ttl_reader::CachedTTLReader;
 
 pub type HybridCacheReader<T, L> = CacheReader<L, HybridCache<T>>;
-pub type InMemoryItemCacheReader<T, L> = CacheReader<L, InMemoryLruCache<T>>;
 pub type InMemoryCacheReader<T, L> = CacheReader<L, InMemoryLruCache<T>>;
-pub type InMemoryCacheTTLReader<T, L> = CachedTTLReader<L, InMemoryLruCache<T>>;
+pub type InMemoryCacheTTLReader<T, L> = CacheTTLReader<L, InMemoryLruCache<T>>;
